requirements_full.txt
appdirs==1.4.4
audioread==2.1.9
backports.zoneinfo==0.2.1
bleak==0.14.3
brython==3.10.5
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
classifier==0.0.0
click==8.0.4
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
docopt==0.6.2
et-xmlfile==1.1.0
ffmpeg==1.4
filelock==3.4.1
Flask==2.2.2
fonttools==4.38.0
gensim==3.7.0
h5py==3.1.0
huggingface-hub==0.12.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
JPype1==1.3.0
Js2Py==0.71
kiwisolver==1.4.4
kobert-transformers==0.4.1
konlpy==0.6.0
librosa==0.6.3
llvmlite==0.31.0
localStoragePy==0.2.3
lxml==4.8.0
MarkupSafe==2.1.2
matplotlib==3.6.3
modules==1.0.0
neosensory-python==0.0.1
nltk==3.6.7
node==0.9.28
nodejs==0.1.1
npm==0.1.1
numba==0.48.0
numexpr==2.8.1
numpy==1.19.5
odict==1.8.0
openpyxl==3.0.9
optional-django==0.1.0
packaging==21.3
pandas==1.0.1
pg==0.1
Pillow==8.4.0
pipreqs==0.4.11
plumber==1.6
pooch==1.5.2
protobuf==3.19.4
psutil==5.9.0
pycparser==2.21
pydub==0.25.1
pygame==2.1.2
pyjsparser==2.7.1
pyobjc-core==8.5
pyobjc-framework-Cocoa==8.5
pyobjc-framework-CoreBluetooth==8.5
pyobjc-framework-libdispatch==8.5
PyOpenGL==3.1.6
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
resampy==0.2.2
sacremoses==0.0.53
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.12.2
sentencepiece==0.1.91
six==1.16.0
smart-open==6.0.0
SoundFile==0.10.3.post1
soynlp==0.0.493
tables==3.7.0
threadpoolctl==3.1.0
tokenizers==0.9.3
torch==1.7.1
tqdm==4.64.0
transformers==3.5.1
typescript==0.0.12
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
Werkzeug==2.2.2
xlrd==2.0.1
yarg==0.1.9
zipp==3.6.0
zope.component==5.0.1
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.lifecycleevent==4.3
